Refactor out new lang_pref djangoapp
