+++ common/djangoapps/lang_pref/__init__.py
@@ -1,0 +1,6 @@
+"""
+Useful information for setting the language preference
+"""
+
+# this is the UserPreference key for the user's preferred language
+LANGUAGE_KEY = 'pref-lang'
