+++ lms/urls.py
@@ -61,6 +61,8 @@ urlpatterns = ('',  # nopep8
 
     url(r'^user_api/', include('user_api.urls')),
 
+    url(r'^lang_pref/', include('lang_pref.urls')),
+
     url(r'^', include('waffle.urls')),
 
     url(r'^i18n/', include('django.conf.urls.i18n')),
